test effectues voir test.txt

+++ create_list.c
@@ -6,7 +6,7 @@
 /*   By: lbaudran <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/17 15:59:18 by lbaudran          #+#    #+#             */
-/*   Updated: 2016/11/17 16:35:12 by lbaudran         ###   ########.fr       */
+/*   Updated: 2016/11/22 11:36:50 by lbaudran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ t_list		*create_elem2(t_list *begin, int champ_nb, int pc)
 {
 	if (!(begin->prev = (t_list *)malloc(sizeof(t_list))))
 		return (NULL);
-	begin->prev->reg_number = ft_memalloc(REG_NUMBER * 4);
+	begin->prev->reg_number = ft_memalloc((size_t)(REG_NUMBER * 4));
 	begin->prev->next = (begin);
 	begin->prev->prev = NULL;
 	begin->prev->pc = pc;
@@ -81,7 +81,7 @@ t_list		*create_elem(t_list *begin, int champ_nb, int pc)
 	{
 		if (!(begin = (t_list *)malloc(sizeof(t_list))))
 			return (NULL);
-		begin->reg_number = ft_memalloc(REG_NUMBER * 4);
+		begin->reg_number = ft_memalloc((size_t)(REG_NUMBER * 4));
 		begin->prev = NULL;
 		begin->next = NULL;
 		begin->pc = 0;
